+++ app/CHANGELOG.md
@@ -23,6 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Delete reported comments
 - Track unsubscribe events
 
+### Fixed
+
+- Redirect after authentication
+
 ## [0.5.0] - 2020-09-22
 
 ### Added